+++ src/main/java/real/world/domain/user/service/UserService.java
@@ -1,4 +1,4 @@
-package real.world.domain.user;
+package real.world.domain.user.service;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
